requirements.txt
streamlit
mysql-connector-python
discord.py